requirements.txt
jsonpath-ng
python-dotenv
python-telegram-bot
requests
